+++ packages/svelte/src/internal/shared/warnings.js
@@ -17,3 +17,22 @@ export function dynamic_void_element_content(tag) {
 		console.warn("dynamic_void_element_content");
 	}
 }
+
+/**
+ * The following properties cannot be cloned with `$state.snapshot` — the return value contains the originals:
+ * 
+ * %properties%
+ * @param {string | undefined | null} [properties]
+ */
+export function state_snapshot_uncloneable(properties) {
+	if (DEV) {
+		console.warn(`%c[svelte] state_snapshot_uncloneable\n%c${properties
+			? `The following properties cannot be cloned with \`$state.snapshot\` — the return value contains the originals:
+
+${properties}`
+			: "Value cannot be cloned with `$state.snapshot` — the original value was returned"}`, bold, normal);
+	} else {
+		// TODO print a link to the documentation
+		console.warn("state_snapshot_uncloneable");
+	}
+}